Minor update for formatting
Set up remote repository on linux.cs.pdx.edu for backup and
cleaned up the file headers a little.

Getting ready to branch to a new NN... each region is a group
of neurons and only one neuron in the group will fire based on
a winner-takes-all for the highest input energy in the group.